backend/requirements.txt
crewai>=0.30.0
qdrant-client>=1.9.1
google-generativeai>=0.5.2
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.1
python-dotenv>=1.0.1
typer>=0.9.0,<0.10.0
streamlit>=1.39.0
plotly>=5.24.0
fpdf2>=2.7.9
pymupdf>=1.24.10
pytesseract>=0.3.13
pillow>=10.4.0